+++ src/hooks/useAuthContext.ts
@@ -1,4 +1,5 @@
 import { useContext } from 'react';
+
 import { AuthContext } from '../context/AuthContext';
 
 export default function useAuthContext() {
