+++ src/hooks/useDeleteAlbum.ts
@@ -1,8 +1,9 @@
-import { appFireStore } from '../firebase/config';
-import { doc, deleteDoc, getDoc } from 'firebase/firestore';
 import { User } from '@firebase/auth';
+import { doc, deleteDoc, getDoc } from 'firebase/firestore';
 
-import useAuthContext from '../hooks/useAuthContext';
+import useAuthContext from './useAuthContext';
+
+import { appFireStore } from '../firebase/config';
 
 export default function useDeleteAlbum() {
   const { user: contextUser } = useAuthContext();
