+++ src/hooks/useEditContext.ts
@@ -1,4 +1,5 @@
 import { useContext } from 'react';
+
 import { EditContext } from '../context/EditContext';
 
 export default function useEditContext() {
