+++ src/hooks/useLogout.ts
@@ -1,4 +1,5 @@
 import { signOut } from 'firebase/auth';
+
 import { appAuth } from '../firebase/config';
 
 export default async function useLogout() {
