+++ src/hooks/usePageContext.ts
@@ -1,4 +1,5 @@
 import { useContext } from 'react';
+
 import { PageContext } from '../context/PageContext';
 
 export default function usePageContext() {
