+++ src/hooks/useUpdateFeedList.ts
@@ -1,6 +1,8 @@
-import { appFireStore } from '../firebase/config';
 import { doc, updateDoc, arrayUnion, arrayRemove } from 'firebase/firestore';
-import useAuthContext from '../hooks/useAuthContext';
+
+import useAuthContext from './useAuthContext';
+
+import { appFireStore } from '../firebase/config';
 
 export function useRemoveFeedIdFromFeedList() {
   const { user } = useAuthContext();
