+++ src/hooks/useUploadImg.ts
@@ -1,5 +1,5 @@
-// 이미지 한 장 업로드
 import { getDownloadURL, ref, uploadBytesResumable } from 'firebase/storage';
+
 import { storage } from '../firebase/config.ts';
 
 export default async function useUploadImg(path: string, file: File) {
